Developer

+++ social-network/backend/java/src/main/java/com/ourproject/socialnetwork/entity/User.java
@@ -14,6 +14,8 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
+
+import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.annotation.Transient;
 import org.springframework.data.mongodb.core.index.Indexed;
@@ -36,7 +38,7 @@ public class User implements UserDetails {
     @Transient
     public static final String SEQUENCE_NAME = "users_sequence";
     @Id
-    private BigInteger userId;
+    private Long userId;
     @Field("user_role")
     @Enumerated(EnumType.STRING)
     private Role role;
